AI-Video-Master_v0.9/requirements.txt
streamlit>=1.22.0
fastapi>=0.85.0
uvicorn>=0.19.0
moviepy>=1.0.3
dashscope>=1.12.0
requests>=2.28.2
python-dotenv>=0.21.0
sentence-transformers>=2.2.2
numpy>=1.22.4
pandas>=1.5.0
plotly>=5.11.0
python-multipart>=0.0.5
Pillow>=9.0.0
nest-asyncio>=1.6.0
